[java][apex] Move *MetricsProvider into own classes into metrics packages
Make *Metrics.reset package private again
