+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/metrics/ApexMetrics.java
@@ -41,7 +41,7 @@ public final class ApexMetrics {
      * Resets the entire data structure.
      * This needs to be done in case PMD is executed multiple times within one JVM run.
      */
-    public static void reset() {
+    static void reset() {
         FACADE.reset();
     }
