+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/JavaMetrics.java
@@ -42,7 +42,7 @@ public final class JavaMetrics {
      * Resets the entire data structure.
      * This needs to be done in case PMD is executed multiple times within one JVM run.
      */
-    public static void reset() {
+    static void reset() {
         FACADE.reset();
     }
